requirements.txt
python-telegram-bot>=13.4
boto3==1.17.101   
botocore==1.20.101
python-dotenv==0.19.0

